Add "Close" pin
This allows us to make sure that we always move the gate in the
correct direction when we want to close it (as we can't be sure what state
it might have been left in).

Need to shift the control button pin over, as 0 should not be held
low on startup.